Feat/server mode

+++ src/tracer/IOTracer.py
@@ -28,6 +28,7 @@ class IOTracer:
             output_dir:         str,
             bpf_file:           str,
             automatic_upload:   bool,
+            server_mode:        bool,
             is_uncompressed:    bool = False,
             anonymous:          bool = False,
             page_cnt:           int = 8,
@@ -46,7 +47,9 @@ class IOTracer:
             if not connection:
                 self.automatic_upload = False
 
-        self.writer             = WriteManager(output_dir, self.upload_manager, automatic_upload)
+        self.server_mode       = server_mode
+
+        self.writer             = WriteManager(output_dir, self.upload_manager, automatic_upload, server_mode)
         self.fs_snapper         = FilesystemSnapper(self.writer, anonymous)
         self.process_snapper    = ProcessSnapper(self.writer, anonymous)
         self.system_snapper     = SystemSnapper(self.writer)
@@ -292,8 +295,9 @@ class IOTracer:
             self.writer.force_flush()
 
             if self.automatic_upload:
-                self.upload_manager.stop_worker()
-                shutil.rmtree(self.writer.output_dir)
+                self.upload_manager.stop_worker(self.server_mode)
+                if self.server_mode:
+                    shutil.rmtree(self.writer.output_dir)
 
             
             logger("info", "Cleanup complete. Exited successfully.")
